+++ MyCodeTestingApp/MyCodeTestingApp/utilities.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <tchar.h>
+
+// These are  tests added for validation
+
+bool _mylogin(const char* kpszUserName, const char* kpszPassword);
+bool _mycheckOverflow(unsigned short x, unsigned short y);
+void _myfoo(int a[10]);
+void _mynewanddelete();
